# docs/swagger.yaml
definitions:
  request.BookPage:
    properties:
      Limit:
        type: integer
      Offset:
        type: integer
      Page:
        type: integer
    type: object
  request.Login:
    properties:
      Password:
        description: 密码
        type: string
      UserId:
        description: 用户名
        type: integer
    type: object
  request.ReqBook:
    properties:
      Author:
        type: string
      Id:
        type: integer
      Introduction:
        type: string
      Name:
        type: string
      Publish:
        type: string
      Statue:
        type: integer
    type: object
  system.Book:
    properties:
      Author:
        description: 作者
        type: string
      AuthorIntroduction:
        description: 作者简介
        type: string
      BookName:
        description: 书名
        type: string
      BriefIntroduction:
        description: 内容简介
        type: string
      DelFlg:
        description: 删除标识
        type: integer
      ISBN:
        description: ISBN号码
        type: string
      Id:
        description: 书的id
        type: integer
      ImgUrl:
        description: 封面地址
        type: string
      Pages:
        description: 页数
        type: integer
      Price:
        description: 价格
        type: number
      PublishDate:
        description: 出版时间
        type: string
      PublishingHouse:
        description: 出版社
        type: string
      Statue:
        type: integer
      Translator:
        description: 译者
        type: string
      UUID:
        type: integer
    type: object
  system.User:
    properties:
      Age:
        type: integer
      Id:
        type: integer
      Identity:
        type: integer
      Name:
        type: string
      Password:
        type: string
      Sex:
        type: string
    type: object
info:
  contact: {}
  description: 测试
  title: 图书管理
  version: 0.0.1
paths:
  /api/public/books:
    get:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.BookPage'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 书
      tags:
      - public
  /api/public/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 用户登录
      tags:
      - public
  /api/public/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 用户注册账号
      tags:
      - public
  /api/public/search:
    get:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: path
        name: BookName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 搜索书
      tags:
      - public
  /api/sys/addbook:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.Book'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 添加图书
      tags:
      - admin
  /api/sys/borrow:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.Book'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 借书
      tags:
      - user
  /api/sys/changeinfo:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.User'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 修改用户信息
      tags:
      - user
  /api/sys/delbook:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.Book'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 删除图书
      tags:
      - admin
  /api/sys/exit:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 退出登录
      tags:
      - user
  /api/sys/information:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 获取用户信息
      tags:
      - user
  /api/sys/records:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 借阅记录
      tags:
      - user
  /api/sys/returnbook:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/system.Book'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 还书
      tags:
      - user
  /api/sys/searchbook:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReqBook'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 搜索某书的借阅记录
      tags:
      - admin
  /api/sys/searchperbook:
    post:
      consumes:
      - application/json
      parameters:
      - description: body参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            type: string
        "500":
          description: 内部错误
          schema:
            type: string
      summary: 搜索某人的借阅记录
      tags:
      - admin
swagger: "2.0"